online_inference/foundatition_models/stable_diffusion_v1_5/requirements.txt
torch
diffusers
transformers
huggingface_hub
pillow
pyyaml
pytest
